removing unecessary .css and new routes file structure
