--- src/components/Root.js
@@ -1,19 +1,0 @@
-import React, { Component } from 'react';
-import PropTypes from 'prop-types';
-import { Provider } from 'react-redux';
-import App from './App';
-
-export default class Root extends Component {
-  render() {
-    const { store } = this.props;
-    return (
-      <Provider store={store}>
-        <App />
-      </Provider>
-    );
-  }
-}
-
-Root.propTypes = {
-  store: PropTypes.object.isRequired
-};
